requirements.txt
Flask~=2.0.3
Flask-JWT-Extended==4.3.1
Flask-RESTful~=0.3.9
Flask-SQLAlchemy==2.5.1
werkzeug==2.0.2